Add files via upload
A simple calculator that takes the current prices of Bitcoin, Litecoin or Etherium and calculates and converts their prices into either American Dollars, British Pounds or Euros. The current value-rates of the crypto currencies are obtained from the API: https://apiv2.bitcoinaverage.com/indices/global/ticker/  

The project has to run locally on the device or on a server, in order to be operational. 

The following npm-modules are required and used in this project:
- express
- bodyParser
- request